% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Activities}
\alias{Activities}
\title{Admin Reports API Objects 
Fetches reports for the administrators of Google Apps customers about the usage, collaboration, security, and risk for their users.}
\usage{
Activities(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{Each record in read response}

\item{nextPageToken}{Token for retrieving the next page}
}
\value{
Activities object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:23:13
filename: /Users/mark/dev/R/autoGoogleAPI/googleadminreportsv1.auto/R/admin_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Activities Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a collection of activites.
}
