% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Activity.actor}
\alias{Activity.actor}
\title{Activity.actor Object}
\usage{
Activity.actor(callerType = NULL, email = NULL, key = NULL,
  profileId = NULL)
}
\arguments{
\item{callerType}{User or OAuth 2LO request}

\item{email}{Email address of the user}

\item{key}{For OAuth 2LO API requests, consumer_key of the requestor}

\item{profileId}{Obfuscated user id of the user}
}
\value{
Activity.actor object
}
\description{
Activity.actor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User doing the action.
}
\seealso{
Other Activity functions: \code{\link{Activity.events.parameters}},
  \code{\link{Activity.events}}, \code{\link{Activity.id}},
  \code{\link{Activity}}
}
