% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Activity.events.parameters}
\alias{Activity.events.parameters}
\title{Activity.events.parameters Object}
\usage{
Activity.events.parameters()
}
\value{
Activity.events.parameters object
}
\description{
Activity.events.parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Activity functions: \code{\link{Activity.actor}},
  \code{\link{Activity.events}}, \code{\link{Activity.id}},
  \code{\link{Activity}}
}
