% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Activity.events}
\alias{Activity.events}
\title{Activity.events Object}
\usage{
Activity.events(Activity.events.parameters = NULL)
}
\arguments{
\item{Activity.events.parameters}{The \link{Activity.events.parameters} object or list of objects}
}
\value{
Activity.events object
}
\description{
Activity.events Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Activity functions: \code{\link{Activity.actor}},
  \code{\link{Activity.events.parameters}},
  \code{\link{Activity.id}}, \code{\link{Activity}}
}
