% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Activity.id}
\alias{Activity.id}
\title{Activity.id Object}
\usage{
Activity.id(applicationName = NULL, customerId = NULL, time = NULL,
  uniqueQualifier = NULL)
}
\arguments{
\item{applicationName}{Application name to which the event belongs}

\item{customerId}{Obfuscated customer ID of the source customer}

\item{time}{Time of occurrence of the activity}

\item{uniqueQualifier}{Unique qualifier if multiple events have the same time}
}
\value{
Activity.id object
}
\description{
Activity.id Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Unique identifier for each activity record.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor}},
  \code{\link{Activity.events.parameters}},
  \code{\link{Activity.events}}, \code{\link{Activity}}
}
