% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\docType{package}
\name{admin_googleAuthR}
\alias{admin_googleAuthR}
\alias{admin_googleAuthR-package}
\title{Admin Reports API
Fetches reports for the administrators of Google Apps customers about the usage, collaboration, security, and risk for their users.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:32:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleadminreportsv1.auto/R/admin_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/admin.reports.audit.readonly
\item https://www.googleapis.com/auth/admin.reports.usage.readonly
}
}

