% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{reports.userUsageReport.get}
\alias{reports.userUsageReport.get}
\title{Retrieves a report which is a collection of properties / statistics for a set of users.}
\usage{
reports.userUsageReport.get(userKey, date, customerId = NULL,
  filters = NULL, maxResults = NULL, pageToken = NULL,
  parameters = NULL)
}
\arguments{
\item{userKey}{Represents the profile id or the user email for which the data should be filtered}

\item{date}{Represents the date in yyyy-mm-dd format for which the data is to be fetched}

\item{customerId}{Represents the customer for which the data is to be fetched}

\item{filters}{Represents the set of filters including parameter operator value}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify next page}

\item{parameters}{Represents the application name, parameter name pairs to fetch in csv as app_name1:param_name1, app_name2:param_name2}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.reports.usage.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.reports.usage.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/reports/}{Google Documentation}
}
