% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReport.entity}
\alias{UsageReport.entity}
\title{UsageReport.entity Object}
\usage{
UsageReport.entity()
}
\value{
UsageReport.entity object
}
\description{
UsageReport.entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the type of the item.
}
\seealso{
Other UsageReport functions: \code{\link{UsageReport.parameters}},
  \code{\link{UsageReport}}
}

