% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReport.parameters}
\alias{UsageReport.parameters}
\title{UsageReport.parameters Object}
\usage{
UsageReport.parameters()
}
\value{
UsageReport.parameters object
}
\description{
UsageReport.parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UsageReport functions: \code{\link{UsageReport.entity}},
  \code{\link{UsageReport}}
}

