% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReport}
\alias{UsageReport}
\title{UsageReport Object}
\usage{
UsageReport(UsageReport.entity = NULL, UsageReport.parameters = NULL,
  etag = NULL)
}
\arguments{
\item{UsageReport.entity}{The \link{UsageReport.entity} object or list of objects}

\item{UsageReport.parameters}{The \link{UsageReport.parameters} object or list of objects}

\item{etag}{ETag of the resource}
}
\value{
UsageReport object
}
\description{
UsageReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a usage report.
}
\seealso{
Other UsageReport functions: \code{\link{UsageReport.entity}},
  \code{\link{UsageReport.parameters}}
}
