% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReports}
\alias{UsageReports}
\title{UsageReports Object}
\usage{
UsageReports(UsageReports.warnings = NULL,
  UsageReports.warnings.data = NULL, etag = NULL, nextPageToken = NULL,
  usageReports = NULL, warnings = NULL)
}
\arguments{
\item{UsageReports.warnings}{The \link{UsageReports.warnings} object or list of objects}

\item{UsageReports.warnings.data}{The \link{UsageReports.warnings.data} object or list of objects}

\item{etag}{ETag of the resource}

\item{nextPageToken}{Token for retrieving the next page}

\item{usageReports}{Various application parameter records}

\item{warnings}{Warnings if any}
}
\value{
UsageReports object
}
\description{
UsageReports Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a collection of usage reports.
}
\seealso{
Other UsageReports functions: \code{\link{UsageReports.warnings.data}},
  \code{\link{UsageReports.warnings}}
}

