% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReports.warnings.data}
\alias{UsageReports.warnings.data}
\title{UsageReports.warnings.data Object}
\usage{
UsageReports.warnings.data()
}
\value{
UsageReports.warnings.data object
}
\description{
UsageReports.warnings.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UsageReports functions: \code{\link{UsageReports.warnings}},
  \code{\link{UsageReports}}
}
