% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UsageReports.warnings}
\alias{UsageReports.warnings}
\title{UsageReports.warnings Object}
\usage{
UsageReports.warnings(UsageReports.warnings.data = NULL)
}
\arguments{
\item{UsageReports.warnings.data}{The \link{UsageReports.warnings.data} object or list of objects}
}
\value{
UsageReports.warnings object
}
\description{
UsageReports.warnings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UsageReports functions: \code{\link{UsageReports.warnings.data}},
  \code{\link{UsageReports}}
}

